+++ CHANGELOG.md
@@ -4,6 +4,7 @@ https://keepachangelog.com/en/1.0.0/
 ## [1.17.2] - 2023-01-02
 - `SynchronousTaskWait`: Now works for top-level statements
 - `SynchronousTaskWait`: Rewritten to use `IOperation`
+- `SynchronousTaskWait`: Preserves leading trivia when applying the code fix
 
 ## [1.17.1] - 2023-01-02
 - `ElementaryMethodsOfTypeInCollectionNotOverridden`: Fixed an issue where a `NullReferenceException` would be thrown